CI: don't test oto & remove libasound2-dev dependency
Installing this dependency failed in the cloud and caused tests to fail, so for now, we do not install it and do not test oto package.
The tests were actually about some float / int16 conversions, which should not anyway be in oto package, so future solution will be to refactor those functions somewhere else.